refactor(memory): move consolidation helpers into package
