+++ packages/ai/memory-consolidation/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./evidence-cluster";
