+++ packages/ai/memory-consolidation/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../config/src/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "rootDir": "./src",
+    "incremental": false,
+    "tsBuildInfoFile": "./dist/.tsbuildinfo",
+    "noEmit": false
+  },
+  "include": ["src"]
+}
