+++ packages/ai/package.json
@@ -26,6 +26,7 @@
     "@vercel/sandbox": "^1.7.1",
     "ai": "5.0.0-beta.6",
     "@ai-sdk/openai-compatible": "^1.0.0",
+    "@openloomi/memory-consolidation": "workspace:*",
     "@openloomi/shared": "workspace:*"
   },
   "peerDependencies": {},
